Implementation of two or more threads merging for multiple platform views (#27662)
Implementation of two or more threads merging for both lightweight multiple engines and standalone multiple engines.

+++ fml/memory/weak_ptr_unittest.cc
@@ -212,14 +212,14 @@ TEST(TaskRunnerAffineWeakPtrTest, ShouldNotCrashIfRunningOnTheSameTaskRunner) {
   fml::TaskQueueId qid2 = loop2->GetTaskRunner()->GetTaskQueueId();
   const auto raster_thread_merger_ =
       fml::MakeRefCounted<fml::RasterThreadMerger>(qid1, qid2);
-  const int kNumFramesMerged = 5;
+  const size_t kNumFramesMerged = 5;
 
   raster_thread_merger_->MergeWithLease(kNumFramesMerged);
 
   loop2_task_start_latch.Signal();
   loop2_task_finish_latch.Wait();
 
-  for (int i = 0; i < kNumFramesMerged; i++) {
+  for (size_t i = 0; i < kNumFramesMerged; i++) {
     ASSERT_TRUE(raster_thread_merger_->IsMerged());
     raster_thread_merger_->DecrementLease();
   }
